backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
google-auth==2.23.0
googlemaps==4.10.0
requests==2.31.0
uuid==1.30
python-slugify==8.0.4
stripe==7.9.0
openai==1.13.3
werkzeug==2.3.7 
flask-socketio==5.3.6
eventlet==0.35.2 